+++ packages/sif-common-forms-ds/src/forms/utenlandsopphold/UtenlandsoppholdForm.tsx
@@ -1,9 +1,14 @@
 import React from 'react';
 import { useIntl } from 'react-intl';
 import FormBlock from '@navikt/sif-common-core-ds/lib/components/form-block/FormBlock';
-import { countryIsMemberOfEøsOrEfta } from '@navikt/sif-common-core-ds/lib/utils/countryUtils';
 import intlHelper from '@navikt/sif-common-core-ds/lib/utils/intlUtils';
-import { DateRange, getCountryName, ISOStringToDate, YesOrNo } from '@navikt/sif-common-formik-ds';
+import {
+    DateRange,
+    getCountryName,
+    countryIsMemberOfEøsOrEfta,
+    ISOStringToDate,
+    YesOrNo,
+} from '@navikt/sif-common-formik-ds';
 import { getTypedFormComponents } from '@navikt/sif-common-formik-ds/lib';
 import {
     getDateRangeValidator,